Move config related code to config package
