+++ compute_endpoint/globus_compute_endpoint/endpoint/config/__init__.py
@@ -1,0 +1,1 @@
+from .config import Config  # noqa: F401
