+++ compute_endpoint/globus_compute_endpoint/endpoint/config/default_config.yaml
