+++ compute_endpoint/tests/integration/endpoint/endpoint/test_interchange_with_rabbit.py
@@ -10,8 +10,8 @@ import pika
 import pytest
 from globus_compute_common.messagepack import pack
 from globus_compute_common.messagepack.message_types import Result, Task
+from globus_compute_endpoint.endpoint.config import Config
 from globus_compute_endpoint.endpoint.interchange import EndpointInterchange
-from globus_compute_endpoint.endpoint.utils.config import Config
 from tests.integration.endpoint.executors.mock_executors import MockExecutor
 from tests.utils import try_for_timeout
