+++ helm/funcx_endpoint/templates/endpoint-instance-config.yaml
@@ -9,7 +9,7 @@ metadata:
     app: {{ .Release.Name }}
 data:
   config.py: |
-    from globus_compute_endpoint.endpoint.utils.config import Config
+    from globus_compute_endpoint.endpoint.config import Config
     from globus_compute_endpoint.executors import HighThroughputExecutor
     from globus_compute_endpoint.providers.kubernetes.kube import KubernetesProvider
     from globus_compute_endpoint.strategies import KubeSimpleStrategy
